Refactor Scalacp to closely follow Metac
Just like Metap, Scalacp has been suffering from its spaghetti structure
for a while now. Hopefully this will untangle things and make Scalacp
better suitable for implementing new functionality.

upd. After the successful first step of the refactoring, I went ahead
and went through AnnotationOps, SymbolInformationOps, SymbolOps and
TypeOps from both Scalacp and Metac in parallel. That resulted in some
nice cleanups.

Note no meaningful changes in Metacp's behavior, except for the fact
that it now emits a proper number of empty `AnnotatedType.annotations`
as opposed to zero `AnnotatedType.annotations` (that's the side effect
of synchronizing Scalacp and Metac).
