+++ semanticdb/metacp/src/main/scala/scala/meta/internal/scalacp/AnnotationOps.scala
@@ -1,0 +1,12 @@
+package scala.meta.internal.scalacp
+
+import scala.meta.internal.{semanticdb3 => s}
+
+trait AnnotationOps { self: Scalacp =>
+  implicit class XtensionAnnotation(ann: Int) {
+    def toSemantic: s.Annotation = {
+      // FIXME: https://github.com/scalameta/scalameta/issues/1292
+      s.Annotation()
+    }
+  }
+}
