+++ semanticdb/scalac/library/src/main/scala/scala/meta/internal/semanticdb/scalac/NameOps.scala
@@ -7,17 +7,11 @@ trait NameOps { self: SemanticdbOps =>
 
   implicit class XtensionName(gname: g.Name) {
     def toSemantic: String = {
-      if (gname == g.nme.ROOTPKG) {
-        n.RootPackage
-      } else if (gname == g.nme.EMPTY_PACKAGE_NAME) {
-        n.EmptyPackage
-      } else if (gname == g.nme.CONSTRUCTOR) {
-        n.Constructor
-      } else if (gname.startsWith("_$")) {
-        n.Anonymous
-      } else {
-        gname.decoded.stripSuffix(g.nme.LOCAL_SUFFIX_STRING)
-      }
+      if (gname == g.nme.ROOTPKG) n.RootPackage
+      else if (gname == g.nme.EMPTY_PACKAGE_NAME) n.EmptyPackage
+      else if (gname == g.nme.CONSTRUCTOR) n.Constructor
+      else if (gname.startsWith("_$")) n.Anonymous
+      else gname.decoded.stripSuffix(g.nme.LOCAL_SUFFIX_STRING)
     }
   }
 }
